+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix find-references and references lens for `defrecord`.
 - Improve `clean-ns` refactoring following the Clojure style guide.
+- Fix `clean-ns` corner case issue where it would replace the `:as` value with the `:refer` value.
 - Deprecate `:keep-require-at-start?` in favor of `:clean :ns-inner-blocks-indentation` with default value of `:next-line`.
 - Bump `org.xerial/sqlite-jdbc` from 3.34.0 to 3.36.0.